refactor: move back web-specific keyboard shortcuts exports into web
