+++ web/src/keyboardShortcuts/KeyboardShortcutsHelp.tsx
@@ -1,7 +1,8 @@
 import { Shortcut } from '@slimsag/react-shortcuts'
 import React, { useCallback, useState } from 'react'
 import { Modal } from 'reactstrap'
-import { KeyboardShortcutsProps, KeyboardShortcut } from '../../../shared/src/keyboardShortcuts'
+import { KeyboardShortcut } from '../../../shared/src/keyboardShortcuts'
+import { KeyboardShortcutsProps } from './keyboardShortcuts'
 
 interface Props extends KeyboardShortcutsProps {
     /** The keyboard shortcut to show this modal. */
