+++ web/src/main.tsx
@@ -15,7 +15,6 @@ import { extensionAreaRoutes } from './extensions/extension/routes'
 import { extensionsAreaHeaderActionButtons } from './extensions/extensionsAreaHeaderActionButtons'
 import { extensionsAreaRoutes } from './extensions/routes'
 import './main.scss'
-import { KEYBOARD_SHORTCUTS } from '../../shared/src/keyboardShortcuts'
 import { orgAreaHeaderNavItems } from './org/area/navitems'
 import { orgAreaRoutes } from './org/area/routes'
 import { repoHeaderActionButtons } from './repo/repoHeaderActionButtons'
@@ -29,6 +28,7 @@ import { userAreaHeaderNavItems } from './user/area/navitems'
 import { userAreaRoutes } from './user/area/routes'
 import { userSettingsAreaRoutes } from './user/settings/routes'
 import { userSettingsSideBarItems } from './user/settings/sidebaritems'
+import { KEYBOARD_SHORTCUTS } from './keyboardShortcuts/keyboardShortcuts'
 
 window.addEventListener('DOMContentLoaded', () => {
     render(
