+++ web/src/nav/GlobalNavbar.tsx
@@ -8,7 +8,6 @@ import * as GQL from '../../../shared/src/graphql/schema'
 import { PlatformContextProps } from '../../../shared/src/platform/context'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { authRequired } from '../auth'
-import { KeyboardShortcutsProps } from '../../../shared/src/keyboardShortcuts'
 import { parseSearchURLQuery, PatternTypeProps } from '../search'
 import { SearchNavbarItem } from '../search/input/SearchNavbarItem'
 import { EventLoggerProps } from '../tracking/eventLogger'
@@ -16,6 +15,7 @@ import { showDotComMarketing } from '../util/features'
 import { NavLinks } from './NavLinks'
 import { ThemeProps } from '../../../shared/src/theme'
 import { ThemePreferenceProps } from '../search/theme'
+import { KeyboardShortcutsProps } from '../keyboardShortcuts/keyboardShortcuts'
 
 interface Props
     extends SettingsCascadeProps,
